cfg/account: Convert the rest of account config parsing.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7459 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -1321,7 +1321,7 @@ struct s_auth
 	int16_t			allowedprotocols;
 	LLIST			*aureader_list;
 	int8_t			autoau;
-	int8_t			monlvl;
+	int32_t			monlvl;
 	uint64_t		grp;
 	int32_t			tosleep;
 	CAIDTAB			ctab;
